refactor: Move CurrentUnitService
