+++ feature-libs/my-account/organization/components/unit/index.ts
@@ -1,7 +1,7 @@
-export * from './current-unit.service';
 export * from './details/index';
 export * from './form/index';
 export * from './guards/index';
 export * from './links/index';
 export * from './list/index';
+export * from './services/current-unit.service';
 export * from './unit-tree/index';
