+++ feature-libs/my-account/organization/components/unit/services/current-unit.service.ts
@@ -2,8 +2,8 @@ import { Injectable } from '@angular/core';
 import { B2BUnit, RoutingService } from '@spartacus/core';
 import { OrgUnitService } from '@spartacus/my-account/organization/core';
 import { Observable } from 'rxjs';
-import { ROUTE_PARAMS } from '../constants';
-import { CurrentOrganizationItemService } from '../shared/current-organization-item.service';
+import { ROUTE_PARAMS } from '../../constants';
+import { CurrentOrganizationItemService } from '../../shared/current-organization-item.service';
 
 @Injectable({
   providedIn: 'root',
