+++ feature-libs/my-account/organization/components/unit/services/unit-item.service.ts
@@ -3,8 +3,8 @@ import { B2BUnit, RoutingService } from '@spartacus/core';
 import { OrgUnitService } from '@spartacus/my-account/organization/core';
 import { Observable } from 'rxjs';
 import { OrganizationItemService } from '../../shared/organization-item.service';
-import { CurrentUnitService } from '../current-unit.service';
 import { UnitFormService } from '../form/unit-form.service';
+import { CurrentUnitService } from './current-unit.service';
 
 @Injectable({
   providedIn: 'root',
